refactor: Increase coverage
