+++ tests/Unit/Traits/CronExpressionValidatorTest.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Stylers\TaskManager\Tests\Unit\Console;
+namespace Stylers\TaskManager\Tests\Unit\Traits;
 
 use InvalidArgumentException;
 use Stylers\TaskManager\Tests\TestCase;
@@ -10,6 +10,7 @@ class CronExpressionValidatorTest extends TestCase
 {
     public function outOfIntervalValuesDataProvider(): array
     {
+        $macroExceptionMsg = 'Macro doesn\'t exists: @non-exists-macro';
         $dayOfWeekExceptionMsg = 'Day of Week must be a number between 0 and 7 (Sunday=0 or 7) or "*", "/", "-", ",".';
         $monthExceptionMsg = 'Month must be a number between 1 and 12 or "*", "/", "-", ",".';
         $dayExceptionMsg = 'Day of Month must be a number between 1 and 31 or "*", "/", "-", ",".';
@@ -17,6 +18,7 @@ class CronExpressionValidatorTest extends TestCase
         $minuteExceptionMsg = 'Minute must be a number between 0 and 59 or "*", "/", "-", ",".';
 
         return [
+            'non-exists-macro' =>           ['@non-exists-macro', $macroExceptionMsg],
             'under_day_of_week_interval' => ['* * * * -1', $dayOfWeekExceptionMsg],
             'above_day_of_week_interval' => ['* * * * 8', $dayOfWeekExceptionMsg],
             'under_month_interval' =>       ['* * * 0 *', $monthExceptionMsg],
